% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agefat}
\alias{agefat}
\title{Age and body fat percentage}
\format{
A data frame with 25 rows and 3 variables.
}
\source{
\pkg{HSAUR3} package.
}
\usage{
agefat
}
\description{
Age, gender, and total body fat percentage (via dual-photon
absorptiometry) for 25 healthy adults. Regression on body fat
percentage from age and gender.
}
\references{
Mazess, R. B., Peppler, W. W., & Gibbons, M. (1984). Total
body composition by dual-photon (153Gd) absorptiometry.
\emph{American Journal of Clinical Nutrition}, 40, 834-839.
}
\keyword{datasets}
