% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aidsdelay}
\alias{aidsdelay}
\title{AIDS reporting delay in England and Wales}
\format{
A data frame with 570 rows and 6 variables.
}
\source{
\pkg{boot} package.
}
\usage{
aidsdelay
}
\description{
Reported AIDS cases in England and Wales (1983-1992), cross-classified
by quarter of diagnosis and reporting delay. Count regression on the
number of cases reported (\code{y}).
}
\references{
De Angelis, D., & Gilks, W. R. (1994). Estimating acquired
immune deficiency syndrome accounting for reporting delay.
\emph{Journal of the Royal Statistical Society, A}, 157, 31-40.
}
\keyword{datasets}
