% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anorexia}
\alias{anorexia}
\title{Anorexia treatment weight change}
\format{
A data frame with 72 rows and 3 variables.
}
\source{
\pkg{MASS} package.
}
\usage{
anorexia
}
\description{
Randomized trial of cognitive behavioural therapy and family therapy
versus control for young female anorexia patients. Regression on
post-treatment weight (\code{Postwt}), adjusting for pre-treatment weight
and treatment arm.
}
\references{
Hand, D. J., Daly, F., McConway, K., Lunn, D., & Ostrowski,
E. (Eds.) (1993). \emph{A Handbook of Small Data Sets}. Chapman &
Hall. Data set 285.
}
\keyword{datasets}
