% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bloodstorage}
\alias{bloodstorage}
\title{Blood storage duration and prostate cancer recurrence}
\format{
A data frame with 316 rows and 20 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
bloodstorage
}
\description{
Retrospective cohort on red blood cell storage duration and biochemical
recurrence of cancer after radical prostatectomy.
}
\references{
Cata, J. P. et al. (2011). Blood storage duration and
biochemical recurrence of cancer after radical prostatectomy.
\emph{Mayo Clinic Proceedings}, 86(2), 120-127.
}
\keyword{datasets}
