% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmtcrr}
\alias{bmtcrr}
\title{Bone marrow transplant competing risks}
\format{
A data frame with 177 rows and 7 variables.
}
\source{
\pkg{casebase} package.
}
\usage{
bmtcrr
}
\description{
Patients who underwent haematopoietic stem cell transplantation for
acute leukemia (ALL/AML), by disease phase and stem-cell source.
Competing-risks survival (relapse vs other failure).
}
\references{
Scrucca, L., Santucci, A., & Aversa, F. (2007). Competing
risk analysis using R: an easy guide for clinicians. \emph{Bone
Marrow Transplantation}, 40(4), 381-387.
}
\keyword{datasets}
