% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breastcancer}
\alias{breastcancer}
\title{Wisconsin breast cancer cytology}
\format{
A data frame with 699 rows and 11 variables.
}
\source{
UCI Machine Learning Repository, via the \pkg{mlbench} package.
}
\usage{
breastcancer
}
\description{
Fine-needle aspirate cytology features for benign vs malignant breast
tumors. Binary classification.
}
\references{
Wolberg, W. H., & Mangasarian, O. L. (1990). Multisurface
method of pattern separation for medical diagnosis applied to breast
cytology. \emph{Proceedings of the National Academy of Sciences},
87, 9193-9196.
}
\keyword{datasets}
