% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burn1000}
\alias{burn1000}
\title{National Burn Repository mortality}
\format{
A data frame with 1000 rows and 9 variables.
}
\source{
\pkg{aplore3} package.
}
\usage{
burn1000
}
\description{
Burn injury patients with total burn surface area, inhalation injury,
and flame involvement, and hospital discharge mortality outcome.
Binary classification.
}
\references{
Hosmer, D. W., Lemeshow, S., & Sturdivant, R. X. (2013).
\emph{Applied Logistic Regression} (3rd ed.). Wiley.
}
\keyword{datasets}
