% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{channing}
\alias{channing}
\title{Channing House retirement community mortality}
\format{
A data frame with 462 rows and 6 variables.
}
\source{
\pkg{KMsurv} package.
}
\usage{
channing
}
\description{
Age at entry and age at death or withdrawal for residents of the
Channing House retirement community. Right-censored survival
(actuarial/demographic cohort).
}
\references{
Hyde, J. (1977). Testing survival under right censoring
and left truncation. \emph{Biometrika}, 64, 225-230.
}
\keyword{datasets}
