% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chdage}
\alias{chdage}
\title{Coronary heart disease and age}
\format{
A data frame with 100 rows and 4 variables.
}
\source{
\pkg{aplore3} package.
}
\usage{
chdage
}
\description{
Classic small dataset relating age to presence of coronary heart
disease. Binary classification.
}
\references{
Hosmer, D. W., Lemeshow, S., & Sturdivant, R. X. (2013).
\emph{Applied Logistic Regression} (3rd ed.). Wiley.
}
\keyword{datasets}
