% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmv}
\alias{cmv}
\title{Donor KIR genotype and CMV reactivation}
\format{
A data frame with 64 rows and 26 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
cmv
}
\description{
Retrospective cohort on donor killer immunoglobulin-like receptor (KIR)
genotype and cytomegalovirus reactivation after allogeneic stem cell
transplant.
}
\references{
Sobecks, R. et al. (2011). Cytomegalovirus reactivation
after matched sibling donor reduced-intensity conditioning allogeneic
hematopoietic stem cell transplant correlates with donor killer
immunoglobulin-like receptor genotype. \emph{Experimental and
Clinical Transplantation}, 1, 7-13.
}
\keyword{datasets}
