% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colon}
\alias{colon}
\title{Colon cancer adjuvant chemotherapy trial}
\format{
A data frame with 1858 rows and 16 variables.
}
\source{
\pkg{survival} package.
}
\usage{
colon
}
\description{
One of the first successful trials of adjuvant chemotherapy for colon
cancer, comparing levamisole and levamisole+fluorouracil. Right-censored
survival (recurrence and death).
}
\references{
Moertel, C. G. et al. (1990). Levamisole and fluorouracil
for adjuvant therapy of resected colon carcinoma. \emph{New England
Journal of Medicine}, 332, 352-358.
}
\keyword{datasets}
