% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covidtest}
\alias{covidtest}
\title{COVID-19 PCR testing cohort}
\format{
A data frame with 15524 rows and 17 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
covidtest
}
\description{
De-identified SARS-CoV-2 PCR testing results and metadata from the
Children's Hospital of Pennsylvania, 2020.
}
\references{
Obstfeld, A. E. (2020). Deidentified results of COVID-19
testing at the Children's Hospital of Pennsylvania (CHOP).
}
\keyword{datasets}
