% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crcmondaca}
\alias{crcmondaca}
\title{Colorectal cancer genomic cohort (Mondaca et al. 2020)}
\format{
A data frame with 471 rows and 19 variables.
}
\source{
\pkg{biostatlab} package.
}
\usage{
crcmondaca
}
\description{
Clinical and tumour-genomic variables (mutation burden, fraction genome
altered, microsatellite status) with right-censored survival outcome.
}
\references{
Mondaca, S. et al. (2020). Colorectal cancer genomic
cohort. Cohort manuscript included with the biostatlab
raw-selected-datasets sources.
}
\keyword{datasets}
