% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug6mp}
\alias{drug6mp}
\title{6-MP leukemia remission trial}
\format{
A data frame with 21 rows and 5 variables.
}
\source{
\pkg{KMsurv} package.
}
\usage{
drug6mp
}
\description{
Freireich et al.'s classic paired clinical trial of 6-mercaptopurine
(6-MP) versus placebo in maintaining remission in acute leukemia. One
of the earliest published randomized survival trials. Right-censored
survival.
}
\references{
Freireich, E. J. et al. (1963). The effect of
6-mercaptopurine on the duration of steroid-induced remissions in
acute leukemia. \emph{Blood}, 21, 699-716.
}
\keyword{datasets}
