% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esophca}
\alias{esophca}
\title{Esophageal cancer case-control study}
\format{
A data frame with 88 rows and 5 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
esophca
}
\description{
Case-control data on alcohol and tobacco use and esophageal cancer.
}
\references{
Breslow, N. E., & Day, N. E. (1980). \emph{Statistical
Methods in Cancer Research. Volume 1: The Analysis of Case-Control
Studies}. IARC Lyon / Oxford University Press.
}
\keyword{datasets}
