% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flchain}
\alias{flchain}
\title{Serum free light chain assay cohort}
\format{
A data frame with 7874 rows and 11 variables.
}
\source{
\pkg{survival} package.
}
\usage{
flchain
}
\description{
Population-based cohort assaying serum free light chain and its
association with overall survival. Right-censored survival.
}
\references{
Dispenzieri, A. et al. (2012). Use of monoclonal serum
immunoglobulin free light chains to predict overall survival in the
general population. \emph{Mayo Clinic Proceedings}, 87, 512-523.
}
\keyword{datasets}
