% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbsg2}
\alias{gbsg2}
\title{German Breast Cancer Study Group 2}
\format{
A data frame with 686 rows and 10 variables.
}
\source{
\pkg{TH.data} package.
}
\usage{
gbsg2
}
\description{
Randomized 2x2 trial evaluating hormonal treatment and duration of
chemotherapy in node-positive breast cancer patients. Right-censored
survival.
}
\references{
Schumacher, M. et al. (1994). Randomized 2x2 trial
evaluating hormonal treatment and the duration of chemotherapy in
node-positive breast cancer patients. \emph{Journal of Clinical
Oncology}, 12, 2086-2093.
}
\keyword{datasets}
