% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glow500}
\alias{glow500}
\title{GLOW osteoporosis fracture risk}
\format{
A data frame with 500 rows and 15 variables.
}
\source{
\pkg{aplore3} package.
}
\usage{
glow500
}
\description{
Global Longitudinal Study of Osteoporosis in Women (GLOW): fracture
risk factors and any-fracture outcome in the first year of follow-up.
Binary classification.
}
\references{
Hosmer, D. W., Lemeshow, S., & Sturdivant, R. X. (2013).
\emph{Applied Logistic Regression} (3rd ed.). Wiley.
}
\keyword{datasets}
