% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haberman}
\alias{haberman}
\title{Haberman breast cancer survival}
\format{
A data frame with 306 rows and 4 variables.
}
\source{
\pkg{biostatlab} package.
}
\usage{
haberman
}
\description{
Survival of patients who underwent surgery for breast cancer, by age,
year of operation, and number of positive axillary nodes. Binary
classification (5-year survival status).
}
\references{
Haberman, S. J. (1976). Generalized residuals for log-linear
models. \emph{Proceedings of the 9th International Biometrics
Conference}, 104-122. UCI Machine Learning Repository.
}
\keyword{datasets}
