% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heartfailure}
\alias{heartfailure}
\title{Heart failure clinical records}
\format{
A data frame with 299 rows and 13 variables.
}
\source{
\pkg{biostatlab} package.
}
\usage{
heartfailure
}
\description{
Clinical records of patients with heart failure, with follow-up time
and death event. Right-censored survival.
}
\references{
Chicco, D., & Jurman, G. (2020). Machine learning can
predict survival of patients with heart failure from serum
creatinine and ejection fraction alone. \emph{BMC Medical Informatics
and Decision Making}, 20, 16.
}
\keyword{datasets}
