% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icu}
\alias{icu}
\title{ICU mortality}
\format{
A data frame with 200 rows and 21 variables.
}
\source{
\pkg{aplore3} package.
}
\usage{
icu
}
\description{
Vital status at hospital discharge for intensive care unit patients,
with admission diagnosis, physiology, and blood-gas covariates.
Binary classification.
}
\references{
Hosmer, D. W., Lemeshow, S., & Sturdivant, R. X. (2013).
\emph{Applied Logistic Regression} (3rd ed.). Wiley.
}
\keyword{datasets}
