% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indometh}
\alias{indometh}
\title{Indomethacin pharmacokinetics}
\format{
A data frame with 66 rows and 3 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
indometh
}
\description{
Cohort study of intravenous indomethacin plasma concentration over
time. Pharmacokinetic regression.
}
\references{
Kwan, K. C., Breault, G. O., Umbenhauer, E. R., McMahon,
F. G., & Duggan, D. E. (1976). Kinetics of indomethacin absorption,
elimination, and enterohepatic circulation in man. \emph{Journal of
Pharmacokinetics and Biopharmaceutics}, 4(3), 255-280.
}
\keyword{datasets}
