% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indorct}
\alias{indorct}
\title{Rectal indomethacin RCT for post-ERCP pancreatitis}
\format{
A data frame with 602 rows and 33 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
indorct
}
\description{
Randomized controlled trial of rectal indomethacin to prevent
pancreatitis after endoscopic retrograde
cholangiopancreatography (ERCP).
}
\references{
Elmunzer, B. J., Higgins, P. D. R., et al. (2012). A
randomized trial of rectal indomethacin to prevent post-ERCP
pancreatitis. \emph{New England Journal of Medicine}, 366, 1414-1422.
}
\keyword{datasets}
