% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidney}
\alias{kidney}
\title{Kidney dialysis catheter infection}
\format{
A data frame with 119 rows and 3 variables.
}
\source{
\pkg{KMsurv} package.
}
\usage{
kidney
}
\description{
Time to first exit-site infection in kidney dialysis patients, by
catheter placement method (surgical vs percutaneous). Right-censored
survival.
}
\references{
Nahman, N. S. et al. (1992). Prevention of catheter-related
infection with exit-site care. \emph{Journal of the American Society
of Nephrology}, 3, 103-107.
}
\keyword{datasets}
