% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laryngoscope}
\alias{laryngoscope}
\title{Video vs standard laryngoscope RCT}
\format{
A data frame with 99 rows and 22 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
laryngoscope
}
\description{
Randomized comparison of the Pentax AWS video laryngoscope versus the
Macintosh laryngoscope in morbidly obese patients.
}
\references{
Abdallah, F. W. et al. (2011). A randomized comparison
between the Pentax AWS video laryngoscope and the Macintosh
laryngoscope in morbidly obese patients. \emph{Anesthesia &
Analgesia}, 113, 1082-1087.
}
\keyword{datasets}
