% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{larynx}
\alias{larynx}
\title{Larynx cancer survival}
\format{
A data frame with 90 rows and 5 variables.
}
\source{
\pkg{KMsurv} package.
}
\usage{
larynx
}
\description{
Survival of larynx cancer patients by disease stage at diagnosis.
Right-censored survival.
}
\references{
Kardaun, O. (1983). Statistical survival analysis of male
larynx-cancer patients. \emph{Statistica Neerlandica}, 37, 103-126.
}
\keyword{datasets}
