% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{licorice}
\alias{licorice}
\title{Licorice gargle RCT for post-extubation cough}
\format{
A data frame with 235 rows and 19 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
licorice
}
\description{
Randomized, double-blind trial of licorice versus sugar-water gargle
before intubation for elective thoracic surgery.
}
\references{
Ruetzler, K. et al. (2013). A randomized, double-blind
comparison of licorice versus sugar-water gargle for prevention of
postoperative sore throat and postextubation coughing. \emph{Anesthesia
& Analgesia}, 117, 614-621.
}
\keyword{datasets}
