% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung}
\alias{lung}
\title{NCCTG lung cancer survival}
\format{
A data frame with 228 rows and 10 variables.
}
\source{
\pkg{survival} package.
}
\usage{
lung
}
\description{
Survival of patients with advanced lung cancer, with performance scores
rated by physician and patient. Right-censored survival.
}
\references{
Loprinzi, C. L. et al. (1994). Prospective evaluation of
prognostic variables from patient-completed questionnaires. North
Central Cancer Treatment Group. \emph{Journal of Clinical Oncology},
12(3), 601-607.
}
\keyword{datasets}
