% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{Malignant melanoma cohort}
\format{
A data frame with 205 rows and 11 variables.
}
\source{
\pkg{riskRegression} package.
}
\usage{
melanoma
}
\description{
Patients with malignant melanoma who underwent radical surgery at
Odense University Hospital, Denmark (1962-1977), followed until 1977.
Competing-risks survival (death from melanoma vs other causes).
}
\references{
Andersen, P. K., & Skovgaard, L. T. (2010). \emph{Regression
with Linear Predictors}. Springer.
}
\keyword{datasets}
