% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabric}
\alias{metabric}
\title{METABRIC breast cancer cohort}
\format{
A data frame with 1904 rows and 693 variables.
}
\source{
\pkg{biostatlab} package.
}
\usage{
metabric
}
\description{
Molecular Taxonomy of Breast Cancer International Consortium cohort:
clinical, gene-expression, and mutation variables with right-censored
overall survival. High-dimensional (n < p).
}
\references{
Curtis, C. et al. (2012). The genomic and transcriptomic
architecture of 2,000 breast tumours reveals novel subgroups.
\emph{Nature}, 486, 346-352. Pereira, B. et al. (2016). The somatic
mutation profiles of 2,433 breast cancers refine their genomic and
transcriptomic landscapes. \emph{Nature Communications}, 7, 11479.
}
\keyword{datasets}
