% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mitrace}
\alias{mitrace}
\title{TRACE myocardial infarction study}
\format{
A data frame with 1877 rows and 9 variables.
}
\source{
\pkg{timereg} package.
}
\usage{
mitrace
}
\description{
Survival after myocardial infarction in relation to clinical heart
failure, diabetes, ventricular fibrillation, and echocardiographic
wall motion index. Right-censored survival.
}
\references{
Jensen, G. V. et al. (1997). Does in-hospital ventricular
fibrillation affect prognosis after myocardial infarction?
\emph{European Heart Journal}, 18, 919-924.
}
\keyword{datasets}
