% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myopia}
\alias{myopia}
\title{Orinda myopia study}
\format{
A data frame with 618 rows and 18 variables.
}
\source{
\pkg{aplore3} package.
}
\usage{
myopia
}
\description{
Longitudinal Study of Myopia: ocular biometry, near-work activity
hours, and parental myopia, with onset of myopia within five years of
follow-up. Binary classification.
}
\references{
Hosmer, D. W., Lemeshow, S., & Sturdivant, R. X. (2013).
\emph{Applied Logistic Regression} (3rd ed.). Wiley.
}
\keyword{datasets}
