% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{NHANES obesity survey extract}
\format{
A data frame with 6482 rows and 21 variables.
}
\source{
\pkg{aplore3} package.
}
\usage{
nhanes
}
\description{
National Health and Nutrition Examination Survey extract: demographic,
cholesterol, blood pressure, and physical-activity variables with
obesity (BMI > 35) outcome. Binary classification.
}
\references{
Hosmer, D. W., Lemeshow, S., & Sturdivant, R. X. (2013).
\emph{Applied Logistic Regression} (3rd ed.). Wiley.
}
\keyword{datasets}
