% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opt}
\alias{opt}
\title{Periodontal therapy and preterm birth}
\format{
A data frame with 823 rows and 171 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
opt
}
\description{
Obstetrics and Periodontal Therapy (OPT) trial data on maternal
periodontal treatment and risk of preterm birth.
}
\references{
Michalowicz, B. S. et al. (2006). Treatment of periodontal
disease and the risk of preterm birth. \emph{New England Journal of
Medicine}, 355, 1885-1894.
}
\keyword{datasets}
