% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paquid}
\alias{paquid}
\title{Paquid dementia cohort}
\format{
A data frame with 2561 rows and 4 variables.
}
\source{
\pkg{riskRegression} package.
}
\usage{
paquid
}
\description{
Prospective French cohort study (PAQUID) on functional and cerebral
ageing, with baseline cognitive test scores. Competing-risks survival
(dementia onset vs death without dementia).
}
\references{
Dartigues, J. F. et al. (1992). The Paquid epidemiological
program on brain ageing. \emph{Neuroepidemiology}, 11(1), 14-18.
}
\keyword{datasets}
