% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbc}
\alias{pbc}
\title{Mayo Clinic primary biliary cholangitis trial}
\format{
A data frame with 418 rows and 20 variables.
}
\source{
\pkg{survival} package.
}
\usage{
pbc
}
\description{
Randomized trial of D-penicillamine for primary biliary cholangitis
(formerly cirrhosis), with follow-up survival data.
}
\references{
Therneau, T., & Grambsch, P. (2000). \emph{Modeling
Survival Data: Extending the Cox Model}. Springer-Verlag, New York.
}
\keyword{datasets}
