% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pima2}
\alias{pima2}
\title{Pima Indians diabetes (corrected)}
\format{
A data frame with 768 rows and 9 variables.
}
\source{
National Institute of Diabetes and Digestive and Kidney
Diseases, via the \pkg{mlbench} package.
}
\usage{
pima2
}
\description{
Diagnostic measurements used to predict diabetes onset in women of Pima
Indian heritage. Corrected version with physiologically impossible zero
values (glucose, blood pressure, triceps, insulin, BMI) recoded as
missing. Binary classification.
}
\references{
Smith, J. W., Everhart, J. E., Dickson, W. C., Knowler,
W. C., & Johannes, R. S. (1988). Using the ADAP learning algorithm to
forecast the onset of diabetes mellitus. \emph{Proceedings of the
Symposium on Computer Applications in Medical Care}, 261-265.
}
\keyword{datasets}
