% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polypharm}
\alias{polypharm}
\title{Polypharmacy in mental health services}
\format{
A data frame with 3500 rows and 14 variables.
}
\source{
\pkg{aplore3} package.
}
\usage{
polypharm
}
\description{
Longitudinal cohort of mental-health service users (long format,
repeated measures per subject across study years), with outcome
polypharmacy (concurrent use of 3+ drug classes). Binary
classification.
}
\references{
Hosmer, D. W., Lemeshow, S., & Sturdivant, R. X. (2013).
\emph{Applied Logistic Regression} (3rd ed.). Wiley.
}
\keyword{datasets}
