% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polyps}
\alias{polyps}
\title{Sulindac RCT for familial adenomatous polyposis}
\format{
A data frame with 22 rows and 7 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
polyps
}
\description{
Randomized trial of sulindac for polyp prevention in familial
adenomatous polyposis.
}
\references{
Giardiello, F. M. et al. (1993). Treatment of colonic and
rectal adenomas with sulindac in familial adenomatous polyposis.
\emph{New England Journal of Medicine}, 328(18), 1313-1316.
}
\keyword{datasets}
