% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{respiratory}
\alias{respiratory}
\title{Respiratory illness clinical trial}
\format{
A data frame with 555 rows and 7 variables.
}
\source{
\pkg{HSAUR3} package.
}
\usage{
respiratory
}
\description{
Longitudinal, multicenter randomized trial comparing an active
treatment to placebo for respiratory illness, with monthly status
assessments (long format, repeated measures per subject). Binary
classification (respiratory status: poor/good).
}
\references{
Davis, C. S. (1991). Semi-parametric and non-parametric
methods for the analysis of repeated measurements with applications
to clinical trials. \emph{Statistics in Medicine}, 10, 1959-1980.
}
\keyword{datasets}
