% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rotterdam}
\alias{rotterdam}
\title{Rotterdam breast cancer cohort}
\format{
A data frame with 2982 rows and 15 variables.
}
\source{
\pkg{survival} package.
}
\usage{
rotterdam
}
\description{
Primary breast cancer patients from the Rotterdam tumor bank, used to
validate the Cox prognostic model. Right-censored survival.
}
\references{
Royston, P., & Altman, D. (2013). External validation of a
Cox prognostic model: principles and methods. \emph{BMC Medical
Research Methodology}, 13, 33.
}
\keyword{datasets}
