% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scurvy}
\alias{scurvy}
\title{James Lind's scurvy trial (1757)}
\format{
A data frame with 12 rows and 8 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
scurvy
}
\description{
Reconstruction of the six-arm treatment comparison from James Lind's
original scurvy trial, one of the first controlled clinical trials.
}
\references{
Lind, J. (1757). \emph{A Treatise on the Scurvy}, in 3
parts, pp. 149-153.
}
\keyword{datasets}
