% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smartpill}
\alias{smartpill}
\title{SmartPill gastric transit cohort}
\format{
A data frame with 95 rows and 22 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
smartpill
}
\description{
Prospective cohort comparing intestinal transit times (via wireless
SmartPill capsule) between trauma patients and healthy volunteers.
}
\references{
Rauch, S. et al. (2012). Use of wireless utility capsule to
determine gastric emptying and small intestinal transit times in
critically ill trauma patients. \emph{Journal of Critical Care},
27(5), 534.e7-534.e12.
}
\keyword{datasets}
