% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streptb}
\alias{streptb}
\title{Streptomycin RCT for pulmonary tuberculosis (1948)}
\format{
A data frame with 107 rows and 13 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
streptb
}
\description{
Reconstruction of the Medical Research Council's streptomycin trial for
pulmonary tuberculosis, widely regarded as the first modern randomized
controlled trial.
}
\references{
Streptomycin in Tuberculosis Trials Committee (1948).
Streptomycin treatment of pulmonary tuberculosis. \emph{British
Medical Journal}, 769-782.
}
\keyword{datasets}
