% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stroke}
\alias{stroke}
\title{Copenhagen Stroke Study}
\format{
A data frame with 518 rows and 15 variables.
}
\source{
\pkg{pec} package.
}
\usage{
stroke
}
\description{
Acute stroke patients with cardiovascular risk factors and stroke
severity score at admission. Right-censored survival.
}
\references{
Jorgensen, H. S., Nakayama, H., Reith, J., Raaschou, H. O.,
& Olsen, T. S. (1996). Acute stroke with atrial fibrillation. The
Copenhagen Stroke Study. \emph{Stroke}, 27(10), 1765-1769.
}
\keyword{datasets}
