% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{support}
\alias{support}
\title{SUPPORT critically ill patients}
\format{
A data frame with 9104 rows and 34 variables.
}
\source{
\pkg{casebase} package.
}
\usage{
support
}
\description{
Study to Understand Prognoses Preferences Outcomes and Risks of
Treatment (SUPPORT): seriously ill hospitalized adults with
demographic, diagnostic, and day-3 physiology covariates.
Right-censored survival, up to 5.56 years of follow-up.
}
\references{
Knaus, W. A., Harrell, F. E., Lynn, J. et al. (1995). The
SUPPORT prognostic model: objective estimates of survival for
seriously ill hospitalized adults. \emph{Annals of Internal
Medicine}, 122, 191-203.
}
\keyword{datasets}
