% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supraclav}
\alias{supraclav}
\title{Supraclavicular nerve block anesthesia study}
\format{
A data frame with 103 rows and 17 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
supraclav
}
\description{
Comparison of combined versus sequential injection of mepivacaine and
ropivacaine for supraclavicular nerve blocks.
}
\references{
Roberman, D. et al. (2011). Combined versus sequential
injection of mepivacaine and ropivacaine for supraclavicular nerve
blocks. \emph{Regional Anesthesia and Pain Medicine}, 36, 145-150.
}
\keyword{datasets}
