% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theoph}
\alias{theoph}
\title{Theophylline pharmacokinetics}
\format{
A data frame with 132 rows and 5 variables.
}
\source{
\pkg{medicaldata} package.
}
\usage{
theoph
}
\description{
Cohort study of oral theophylline plasma concentration over time.
Pharmacokinetic regression.
}
\references{
Boeckmann, A. J., Sheiner, L. B., & Beal, S. L. (1994).
\emph{NONMEM Users Guide: Part V}. NONMEM Project Group, University
of California, San Francisco.
}
\keyword{datasets}
