% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toenail}
\alias{toenail}
\title{Toenail infection clinical trial}
\format{
A data frame with 1908 rows and 5 variables.
}
\source{
\pkg{HSAUR3} package.
}
\usage{
toenail
}
\description{
Randomized trial comparing two oral antifungal treatments for toenail
infection (dermatophyte onychomycosis), with outcome assessed at up to
seven visits (long format, repeated measures per subject). Binary
classification (degree of nail separation).
}
\references{
De Backer, M. et al. (1998). Twelve weeks of continuous
oral therapy for toenail onychomycosis caused by dermatophytes.
\emph{Journal of the American Academy of Dermatology}, 38, S57-S63.
}
\keyword{datasets}
