% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{veteran}
\alias{veteran}
\title{Veterans' Administration lung cancer trial}
\format{
A data frame with 137 rows and 8 variables.
}
\source{
\pkg{survival} package.
}
\usage{
veteran
}
\description{
Randomized trial comparing two treatment regimens for lung cancer.
Right-censored survival.
}
\references{
Kalbfleisch, D., & Prentice, R. L. (1980). \emph{The
Statistical Analysis of Failure Time Data}. Wiley, New York.
}
\keyword{datasets}
